Extracted count candidate and hover data
