+++ decoder/body_candidates.go
@@ -146,17 +146,3 @@ func isBlockDeclarable(body *hclsyntax.Body, blockType string, bSchema *schema.B
 	}
 	return true
 }
-
-func countAttributeCandidate(editRng hcl.Range) lang.Candidate {
-	return lang.Candidate{
-		Label:       "count",
-		Detail:      "optional, number",
-		Description: lang.PlainText("The distinct index number (starting with 0) corresponding to the instance"),
-		Kind:        lang.AttributeCandidateKind,
-		TextEdit: lang.TextEdit{
-			NewText: "count",
-			Snippet: "count = ${1:1}",
-			Range:   editRng,
-		},
-	}
-}
